requirements-dev.txt
# Development and testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
isort>=5.12.0

# Testing utilities
requests-mock>=1.10.0
freezegun>=1.2.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0

# Development tools
ipython>=8.0.0
jupyter>=1.0.0